services/oca3/requirements.txt
connexion[swagger-ui]>= 2.7.0
swagger-ui-bundle >= 0.0.8
python_dateutil >= 2.8.1
setuptools >= 50.0.3
google-cloud-firestore==1.9.0
google-cloud-logging==1.15.1
